# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API permite generar y validar códigos QR para invitaciones a eventos.",
        "title": "API de Generación y Validación de Códigos QR",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/qr-code",
    "paths": {
        "/generate": {
            "post": {
                "description": "Genera un código QR basado en la información proporcionada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QR"
                ],
                "summary": "Generar código QR",
                "parameters": [
                    {
                        "description": "Datos para generar el código QR",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.QRManagement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Los códigos QR se están generando",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            }
        },
        "/generate_batch": {
            "post": {
                "description": "Genera múltiples códigos QR basados en la información proporcionada en un archivo CSV",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QR"
                ],
                "summary": "Generar códigos QR en lote",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Archivo CSV con los datos de los invitados | Formato CSV (nombre, invitado_por, parentesco)",
                        "name": "Invitaciones",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Los códigos QR se están generando",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            }
        },
        "/validate/{id}": {
            "get": {
                "description": "Valida un código QR basado en el ID proporcionado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QR"
                ],
                "summary": "Validar código QR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del código QR",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QRManagement"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.QRManagement": {
            "type": "object",
            "properties": {
                "creado": {
                    "type": "string",
                    "example": ""
                },
                "entregado": {
                    "type": "string",
                    "example": ""
                },
                "invitado_por": {
                    "type": "string",
                    "example": "Neifer"
                },
                "nombre": {
                    "type": "string",
                    "example": "Pedro Pacheco"
                },
                "parentesco": {
                    "type": "string",
                    "example": "Familia"
                },
                "sorteo": {
                    "type": "string",
                    "example": ""
                },
                "status": {
                    "type": "string",
                    "example": "Created"
                }
            }
        },
        "entity.Error": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "entity.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}